:sparkles: domain Workflow and worker bootstrap
A Workflow is just, for now, a function that could be run through a remote trigger.
It is defined by a `name` to be referenced.
A Worker allows to run a Workflow on a specific environment.
The trigger that starts the workflow and the Worker that actually execute the specified workflow could be one difference processes. An EventBus is used to communicate between this 2 processes.
In this first implementation, a simple in-memory eventbus is implemented in order to test the flow.